==> scripts/00_Configuration_Update.ipynb <==
### Script to set initial values, reset or updates values within configuration files

#############################################
# Prerequisites - Load Libraries

source('MS1_Functions.r')

### Notify about start of execution
popup_function_pos('00_Configuration_Update: Execution Started')

source('MS0_Libraries.r')

source('MS2_Plot_Config.r')

###############################################
# Preqrequisites Configurations & Parameters

### Load the parameters that are set via the configuration files

### Load configurations file
global_configs = read.csv('configurations/Data_Configs.csv', sep = ',')

head(global_configs,3)

### Path to the data to be used within the analysis

data_path = global_configs$value[global_configs$parameter == 'data_path']

data_path

### Path to the folder where results of the analysis will be stored

result_path = global_configs$value[global_configs$parameter == 'result_path']

result_path

### Default configuration name added to all configuration files; can also be adjusted per configuration file

configuration_name = global_configs$value[global_configs$parameter == 'configuration_name']

configuration_name

# Adjust values in existing configuration files

### List files in input data folder

available_files = list.files(data_path)

available_files

### Keep only 'h5ad' or '.csv' files

available_files = available_files[!is.na(str_extract(available_files, '.h5ad|.csv'))] 

available_files

## 01_Pre_Processing_SC_Data

### Add names of sc dataset to '01_Pre_Processing_SC_Dataset' (all h5ad datasets)
if(sum( !is.na(str_extract(available_files, '.h5ad'))) >=1){
    data = data.frame('data_name' = str_replace(available_files[!is.na(str_extract(available_files, '.h5ad'))], '.h5ad|.csv', ''))
    write.csv(data, 'configurations/01_Pre_Processing_SC_Data.csv', row.names = FALSE)
                      }

## 02_Pre_Processing_Configs_SC

configs02_sc = data.frame(
    configuration_name = configuration_name,
    data_name = str_replace(available_files[!is.na(str_extract(available_files, '.h5ad'))], '.h5ad|.csv', ''),
    data_type = str_extract(available_files[!is.na(str_extract(available_files, '.h5ad'))], '.h5ad|.csv'),
    cell_expr_thres1 = '50;10',
    cell_expr_thres2 = '40;20',
    cell_type_exclusion = ''
)
    

configs02_sc

write.csv(configs02_sc, 'configurations/02_Pre_Processing_Configs_SC.csv', row.names = FALSE)

## 02_Pre_Processing_Configs

### Add names of other datasets to '02_Pre_Processing_Configs'

configs02 = data.frame(
    configuration_name = configuration_name,
    data_name = str_replace(available_files, '.h5ad|.csv', ''),
    file_type = str_extract(available_files, '.h5ad|.csv'),
    data_type = str_replace(available_files, '.h5ad|.csv', ''),
    remove_sample_ids = '',
    sample_filtering_thres = 1,
    feature_filtering_thres = 0.05,
    library_adjustment = FALSE,
    log_transformation = TRUE,
    variable_genes_filtering = 1,
    quantile_normalization_samples = FALSE,
    ribosomal_mitochondrial_gene_filtering = FALSE,
    feature_wise_quantile_normalization = TRUE)
    
    
    
    
    
    

configs02

write.csv(configs02, 'configurations/02_Pre_Processing_Configs.csv', row.names = FALSE)

## 03_MOFA_Configs

### Adjust 03 MOFA configs

configs03 = data.frame(
    configuration_name = configuration_name,
    mofa_result_name = paste0(configuration_name, '_MOFA'),
    amount_of_factors = 10,
    weighting_of_views = FALSE,
    scale_views = TRUE)
    

configs03

write.csv(configs03, 'configurations/03_MOFA_Configs.csv', row.names = FALSE)

## 04_Factor_Analysis_Configs

### Adjust 04 Factor Analysis

configs04 = data.frame(
    configuration_name = configuration_name,
    mofa_result_name = paste0(configuration_name, '_MOFA'),
    relevant_factors = c("Factor1,Factor2,Factor3,Factor4,Factor5,Factor6"),
    numeric_covariates = '',
    categorical_covariates = '',
    top_variable_thres  = 0.005)

configs04

write.csv(configs04, 'configurations/04_Factor_Analysis_Configs.csv', row.names = FALSE)

## 05_Feature_Analysis_Configs

### Adjust 05 Feature Analysis Configs

configs05 = data.frame(
    configuration_name = configuration_name,
    mofa_result_name = paste0(configuration_name, '_MOFA'),
    factor = 'Factor1',
    top_variable_thres = 0.005,
    faceting_variable = '',
    type = '',
    plot_width =8.07 ,
    plot_height = 5)
    

write.csv(configs05, 'configurations/05_Feature_Analysis_Configs.csv', row.names = FALSE)

## 06_Pathway_Configs

### Adjust 06 Pathway Analysis Configs

configs06 = data.frame(
    mofa_result_name = paste0(configuration_name, '_MOFA'),
    factor_set = '1',
    coverage_par = 0.2,
    types = '',
    coverage_plot  = 0.5,
    p_value_plot = 0.05,
    max_pathways_plot = 8,
    enrichment_plot = 'positive',
    top_features_plot = 0.125,
    pathway_selection = '')

configs06

write.csv(configs06, 'configurations/06_Pathway_Configs.csv', row.names = FALSE)

## 07_Comparison_Configs

## Adjust 07 comparison configs file

configs07 = data.frame(
    mofa_result_name = paste0(configuration_name, '_MOFA'),
    compare_factors = 'Factor1,Factor2,Factor3')

configs07

write.csv(configs07, 'configurations/07_Comparison_Configs.csv', row.names = FALSE)

### Notify that exeuction finished successfully

popup_function_pos('00_Configuration_Update: Execution Finished')

Sys.sleep(20)
popup_function_info('00_Configuration_Update')
