==> scripts/03_Run_MOFA.ipynb <==
### Run MOFA Model on the data generated in the previous script (02)

#############################################
# Prerequisites - Load Libraries

source('MS1_Functions.r')

### Informa about start of execution

popup_function_pos('03_Run_MOFA: Execution Started')

source('MS0_Libraries.r')

source('MS2_Plot_Config.r')

#py_config() # - To check the configuration which python package will be used for MOFA

###############################################
# Preqrequisites Configurations & Parameters

### Load the parameters that are set via the configuration files

### Load configurations file
global_configs = read.csv('configurations/Data_Configs.csv', sep = ',')

head(global_configs,2)

data_path = global_configs$value[global_configs$parameter == 'data_path']

data_path

result_path = global_configs$value[global_configs$parameter == 'result_path']

result_path

## MOFA Model Configurations

mofa_configs = read.csv( 'configurations/03_MOFA_Configs.csv', sep = ',')
mofa_configs = mofa_configs[mofa_configs$configuration_name != '',]

head(mofa_configs,2)

### Generate the result data directory if it does not exist yet
if(!file.exists(paste0(result_path, '03_results'))){
    dir.create(file.path(paste0(result_path, '03_results')))
    }

# Load Data 

## Prepared combined data

### Load the data that was generated in the previous script using the name specified in the configuration file

input_data = list()

for(i in 1:nrow(mofa_configs)){
    path = paste0(result_path, '/02_results/02_Combined_Data_', mofa_configs$configuration_name[i] ,'_INTEGRATED.csv')
    
    if(file.exists(path)){
        data_long = read.csv(path)
        data_long$X = NULL
        print(path)
        print(file.info(path)$mtime)
        input_data[[i]]= data_long
        popup_function_pos('Prepared Data has been loaded')
        }
    else{
        popup_function_neg(paste0( 'Error: Data at' , result_path, '/02_results/', ' 02_Combined_Data_', mofa_configs$configuration_name[i] ,'_INTEGRATED.csv', ' could not be loaded. Make sure that all the previous steps have been executed successfully'))
     }   
        
    }

length(unique(input_data[[1]]$variable))

unique(input_data[[1]]$type)

length(unique(input_data[[1]]$sample_id))

# Train MOFA Model

## Prepare data list

### Adjust single-cell types to correspond to cell-types

head(input_data[[1]],2)

### Prepare data list for MOFA (adjust format of input data to be used as input for MOFA)

input_data_list = list()

data_list = list()

input_data_list = lapply(input_data, function(x){

    for(i in unique(x$type)){
        samples = unique(x$sample_id) # necessary to have all samples in all dimensions
        data = x[x$type == i, ]

        data$type = NULL
        data$cell_type = NULL

        data = data %>% dcast(variable ~ sample_id, value  = "value")
        rownames(data) = data$variable
        colnames(data) = str_replace(colnames(data), 'value\\.', '')
        data$variable = NULL

        data[setdiff( samples, names(data))] = NA  # use all samples

        data = data[,order(colnames(data))]
        data = data[,colnames(data) %in% samples]

        data_list[[i]] = as.matrix(data)
        }
    
    return(data_list)
    })

head(input_data_list[[1]][[1]],2)

#str(input_data_list)

## Create MOFA object

### Create a MOFA object to run the MOFA model on it

names(input_data_list[[1]])

mofa_object = lapply(input_data_list, function(x){
    MOFAobject = create_mofa(x)
    }
                     )

### Plot the Data Overview showing the input used for the MOFA Model

# Specific Text Descriptions:
xlabel = xlab('Samples') 
ylabel = ylab('View')

# Sizes of the plot
width_par = 5
height_par =5

options(repr.plot.width=30, repr.plot.height=10)

mofa_overview = lapply(mofa_object, function(x){
    mofa_overview = plot_data_overview(x)
    mofa_overview = mofa_overview + plot_config + theme(axis.text.y = element_text(hjust = 0, vjust = 0.5)) +
                xlabel + ylabel + theme(axis.text.x = element_blank())
    })

#mofa_overview[[1]]

# Extract data -type colors (used by the function to align and use those colors in the next plots)
type_colors = list()
for(i in 1:length(mofa_overview)){
    color_extraction =  ggplot_build(mofa_overview[[i]])
    type_colors[[i]] = unique(color_extraction$data[[1]]["fill"][,1])
    type_colors[[i]] = type_colors[[i]][!type_colors[[i]] == 'grey']
    }
    

type_colors

figure_name = "FIG03_Overview_MOFA_Input_"


for(i in 1:length(mofa_overview)){
    pdf(paste0('figures/03_figures/', figure_name, mofa_configs$mofa_result[i],  '.pdf'), width =width_par, height =height_par)
    print(mofa_overview[[i]] )
    dev.off()
    }
popup_function_pos(paste0('Generated: figures/03_figures/ ', figure_name, mofa_configs$mofa_result[i],  '.pdf'))

## Set MOFA Training Options and run the Model Training

### Define the MOFA parameters for training and run the model training with the set parameters
### Some parameters are handed over by thhe configuration file
### Others are currently assigned fixed below but can be modified

model_result = list()

for(i in 1:length(mofa_object)){
    
    ## Set other parameters of MOFA Model
    mefisto_opts = get_default_mefisto_options(mofa_object[[i]])
    
    ## Data Options
    data_opts = get_default_data_options(mofa_object[[i]])
    data_opts$scale_views = mofa_configs$scale_views[i] # decide whether to scale the data
    data_opts$use_float32 = FALSE
    print(data_opts)
    
    ## Model Options
    model_opts = get_default_model_options(mofa_object[[i]])
    model_opts$num_factors = mofa_configs$amount_of_factors[i] # define number of factors
    model_opts$spikeslab_weights = TRUE
    # model_opts$likelihoods['neutrophil'] = 'poisson' - example to modify distribution for one specific view
    print(model_opts)
    
    ## Training Options
    train_opts  = get_default_training_options(mofa_object[[i]])
    train_opts$maxiter = 50000
    train_opts$verbose = TRUE
    train_opts$seed = 42
    train_opts$weight_views = mofa_configs$weighting_of_views[i]
    print(train_opts)
    
    ## Build and train the model
    MOFAobject = prepare_mofa(
      object = mofa_object[[i]],
      data_options = data_opts,
      model_options = model_opts,
      mefisto_options = mefisto_opts,
      training_options = train_opts #,
      #stochastic_options = stoch_options
    )
    
    model_name = paste0("03_MOFA_MODEL_", mofa_configs$mofa_result[i], '.hdf5')
    outfile = file.path( paste0(result_path, '/03_results/',  model_name) )
    print(outfile)
    MOFAobject.trained = run_mofa(MOFAobject, outfile, use_basilisk = FALSE)
    

    model_result[[i]] = MOFAobject.trained
    
    }
    

#str(model_result)

reticulate::py_config()

# Extract and prepare data for plots

### Extract generated data for the model to use for later downstream analysis

## Extract Variance decomposition

# Extract the total explained variance per view and factor

model_result[[1]]@cache[["variance_explained"]]$r2_total  # per view

rowMeans(model_result[[1]]@cache$variance_explained$r2_per_factor[[1]]) # per factor

# Mean total variance explained

mean(model_result[[1]]@cache$variance_explained$r2_total[[1]])

# Save the explained variance

for(i in 1:length(model_result)){
    write.csv(model_result[[i]]@cache$variance_explained$r2_per_factor[[1]], paste0(result_path, '/03_results/03_MOFA_Variance_Decomposition_',mofa_configs$mofa_result[i], '.csv'))
    }
popup_function_pos(paste0('Saved:', result_path, ' /03_results/ 03_MOFA_Variance_Decomposition_',mofa_configs$mofa_result[i], '.csv'))

## Extract factor and weight data

#### Extract sample factors  values and save

for(i in 1:length(model_result)){
    factors = get_factors(model_result[[i]], factors = "all")
    factors = factors$group1
    head(factors,2)
    
    factors = as.data.frame(factors)
    factors$sample_id = rownames(factors)
    
    # Save as csv
    write.csv(factors, paste0(result_path, '/03_results/03_Factor_Data_' , mofa_configs$mofa_result[i],  '.csv'), row.names = FALSE)
    }
popup_function_pos(paste0('Saved:', result_path, ' /03_results/ 03_Factor_Data_' , mofa_configs$mofa_result[i],  '.csv'))

### Extract weight data (feature factor weights) and save

for(j in 1:length(model_result)){
    weights = get_weights(model_result[[j]], views = "all", factors = "all")
    weight_data = data.frame()
    
    for (i in names(weights)){
        data = data.frame(weights[[i]])
        data$type = i
        weight_data = rbind(weight_data,data)
        }
    weight_data$variable_name = rownames(weight_data)
    
    # Save as csv
    write.csv(weight_data, paste0(result_path, '/03_results/03_Weight_Data_' ,mofa_configs$mofa_result[j], '.csv'), row.names = FALSE)
    }
popup_function_pos(paste0('Saved:', result_path, ' /03_results/ 03_Weight_Data_' ,mofa_configs$mofa_result[j], '.csv'))

# Diagnostic Result Plots

### Make the explained variance plot to analyze the model result

## Plot explained variance overview

## Prepare the data format

explained_variance = lapply(model_result, function(x) {
    data = x@cache$variance_explained$r2_per_factor[[1]]
    data = melt(data)
    
    total_variance = data.frame( view = rownames(x@cache[["variance_explained"]]$r2_total$group1,2),
                             total_variance = x@cache[["variance_explained"]]$r2_total$group1)
    data = merge(data, total_variance, by.x = 'Var2', by.y = 'view')
    data$Var2 = as.character(data$Var2)
    data$Var2 = factor(data$Var2, levels = sort(unique(data$Var2)))
    data = data[order(data$Var2),]
    }
                            )

#### Plot complete explained variance (Heatmap)

var_decomp = lapply(explained_variance, function(x){
    ggplot() + 
        scale_fill_gradient(low="white", high="black") + 
        xlabel + 
        ylabel +
        plot_config + theme(axis.text.x = element_text(angle = 90), legend.position = 'right')+ 
        geom_tile(data = x, mapping = aes(Var1,  Var2, fill= value))
    })

### Combine the plot with total variance barplot per dimension

# Specific Text Descriptions:
xlabel = xlab('View') 
ylabel = ylab('Explained Variance')

comp_variance = lapply(explained_variance, function(x){
    data = x
    plot_complete = unique(data[,c('Var2', 'total_variance')])
    comp_variance = ggplot(plot_complete, aes(x=Var2, y = total_variance, fill = Var2)) + 
                geom_bar(stat="identity") + coord_flip() + 
                xlabel + 
                ylabel +
                plot_config + scale_fill_manual(values = unlist(type_colors))  ## currently uses same coloring as MOFA oveview
    })

#comp_variance[[1]]

### Combine both visualization

figure_name = "FIG03_Overview_Variance_Decomposition"

# Sizes of the plot
width_par = 8.07
height_par = 4  # 2.8

for(i in 1:length(explained_variance)){
    legend = get_legend(var_decomp[[i]])
    
    combination1 = ggarrange(var_decomp[[i]] + theme(legend.position = 'none'),
                     comp_variance[[i]] + theme(axis.text.y = element_blank(),axis.ticks.y = element_blank(),axis.title.y = element_blank(), legend.position = 'none' ), 
                         align = 'h', nrow=1, widths = c(4,1))
    # Annotate Figure
    combination1_ann = annotate_figure(
      combination1,
      right = legend
    )
    
    pdf(paste0('figures/03_figures/', figure_name,  mofa_configs$mofa_result[i],  '.pdf'), width =width_par, height =height_par)
    print(combination1_ann)
    dev.off()
    #print(combination1_ann)
    
    }
popup_function_pos(paste0('Generated:', ' figures/03_figures/ ', figure_name,  mofa_configs$mofa_result[i],  '.pdf'))  

## Save view colors for further usage

write.csv(data.frame(color_code = unlist(type_colors)),
          paste0('configurations/03_Type_Color_Codes.csv'))

## Inform that execution was finished
popup_function_pos('03_Run_MOFA: Execution Finished')

Sys.sleep(20)
popup_function_info('03_Run_MOFA')
