==> scripts/05_Downstream_Investigate_Features_Heatmap.ipynb <==
### Analyze MOFA MODEL Results
### Investigate the specific top features of each factor
### Plot the normalized expression of the features as heatmap

#############################################
# Prerequisites - Load Libraries

source('MS1_Functions.r')

### Inform about execution start
popup_function_pos('05_Downstream_Investigate_Features_Heatmap: Execution Started')

source('MS0_Libraries.r')

source('MS2_Plot_Config.r')

###############################################
# Preqrequisites Configurations & Parameters

### Load the parameters that are set via the configuration files

### Load configurations file
global_configs = read.csv('configurations/Data_Configs.csv', sep = ',')

head(global_configs,2)

data_path = global_configs$value[global_configs$parameter == 'data_path']

data_path

result_path = global_configs$value[global_configs$parameter == 'result_path']

result_path

### Configuration parameters for plot (read and assign)

plot_configs =  read.csv('configurations/05_Feature_Analysis_Configs.csv', sep = ',')
plot_configs = plot_configs[plot_configs$configuration_name != '',]

head(plot_configs,2)

## Assign the values from the configuration file to the corresponding variables

mofa_name = plot_configs$mofa_result_name
data_name = plot_configs$configuration_name

factor_var = plot_configs$factor
top_variable_fraction = plot_configs$top_variable_thres
variable = plot_configs$faceting_variable

view_var = plot_configs$type

## Extract the specified plot size
width_par = plot_configs$plot_width  # 8.07 default
height_par = plot_configs$plot_height

### Alternative: Set configurations manually 

# Load Data 

## MOFA Input

### Load the normalized expression data that was used as input for the MOFA model

path = paste0(result_path, '/02_results/02_Combined_Data_',data_name,'_INTEGRATED.csv')
if(file.exists(path)){
    data_long = read.csv(path)
    data_long$X = NULL
    print(file.info(path)$mtime)
    popup_function_pos(paste0('Loaded: ', '02_Combined_Data_',data_name,'_INTEGRATED.csv'))
    
        } else { popup_function_neg(paste0('Error: Cannot Load: ', '02_Combined_Data_',data_name,'_INTEGRATED.csv', ' Please check whether the previous step has been executed successfully'))} 


head(data_long,2)

## MOFA Model

### Load the trained MOFA Model

model_name =  paste0("03_MOFA_MODEL_", mofa_name,'.hdf5')

outfile = file.path( paste0(result_path, '/03_results/',  model_name) )

outfile

if(file.exists(outfile)){
    model <- load_model(outfile, verbose = TRUE)
    popup_function_pos(paste0('Loaded: ', model_name))
    } else {popup_function_neg(paste0('Error: ',result_path, '/03_results/ ', model_name, 'could not be loaded. Check whether the previous steps have been executed successfully')) } 

### Load the Factor Data of the Model

path = paste0(result_path, '/03_results/', '03_Factor_Data_', mofa_name, '.csv')

if(file.exists(path)){
    factors = read.csv(path)
    popup_function_pos(paste0('Loaded: ','03_Factor_Data_', mofa_name, '.csv'))
    } else {popup_function_neg(paste0('Error: ','03_Factor_Data_', mofa_name, '.csv', ' could not be loaded. Check whether the previous steps have been executed successfully'))}
    

head(factors,2)

### Load the Weight Data of the Model

path = paste0(result_path, '/03_results/', '03_Weight_Data_', mofa_name, '.csv')
if(file.exists(path)){
    weight_data = read.csv(path)
    popup_function_pos(paste0('Loaded: ','03_Weight_Data_',  mofa_name, '.csv'))} else{ popup_function_neg(paste0('Error: ','03_Weight_Data_',  mofa_name, '.csv', ' could not be loaded. Check whether the previous steps have been executed successfully and the file exists.'))} 

## Sample Meta Data

## Load the sample meta-data that contains covariates that can be used to group the plot

file_path = paste0(data_path, 'Prepared_Sample_Meta_Data', '.csv')
if(file.exists(file_path)){
    sample_data = read.csv(paste0(data_path, 'Prepared_Sample_Meta_Data', '.csv'))
    sample_data$X = NULL
    popup_function_pos(paste0('Loaded: ','Prepared_Sample_Meta_Data', '.csv'))} else{ popup_function_neg(paste0('Error: ','Prepared_Sample_Meta_Data', '.csv', ' could not be loaded. Please add the file to the specified input data folder.'))}

head(sample_data,2)

## Check whether faceting variable is included in sample-meta-data file

if (!plot_configs$faceting_variable %in% colnames(sample_data)){
    popup_function_neg(paste0('Error: ', plot_configs$faceting_variable, ' is not included in the uploaded sample data file. Specify another faceting variable or include it in the sample data file'))
    print(paste0('Error: ', plot_configs$faceting_variable, ' is not included in the uploaded sample data file. Specify another faceting variable or include it in the sample data file'))

}
    

## Check whether the sample_id column is included

if(! 'sample_id' %in% colnames(sample_data)){
    popup_function_neg(paste0('Error: ', 'columns sample_id is not included in sample data and is required'))
    print(paste0('Error: ', 'columns sample_id is not included in sample data and is required'))
}

# Downstream Analysis of generated model

## Extract and prepare data for plots

### Merge factors and sample data

### Combine the factor and the sample data

head(factors,2)

factors_long = melt(factors, id.vars = 'sample_id')

head(factors_long,2)

merged_data_long = merge(factors_long, sample_data, by.x = 'sample_id', by.y = 'sample_id')

head(merged_data_long,2)

### Prepare weight data

### Adjust the format of the feature factor weight data to long

feature_weights_long = melt(weight_data, id.vars = c('variable_name', 'type'))

feature_weights_long$view = feature_weights_long$type
feature_weights_long$gene = str_replace(feature_weights_long$variable_name, '.*__', '')

head(feature_weights_long,2)

### Get top features per factor and for specified threshold

## Get the x% of top features per factor based on the specified threshold (x) in the configuration file

dimensions = unique(feature_weights_long[,c('view', 'variable')])


## Define amont of top genes per fraction 
geneset_oi_pos_per_factor_analyze = feature_weights_long %>% group_by(variable) %>% dplyr::arrange( desc(value),  .by_group = TRUE)  %>% top_frac( as.numeric(top_variable_fraction), value)
geneset_oi_pos_per_factor_analyze$direction = 'positive'

geneset_oi_neg_per_factor_analyze = feature_weights_long %>% group_by(variable) %>% dplyr::arrange(desc(value),  .by_group = TRUE)  %>% top_frac( -as.numeric(top_variable_fraction), value)
geneset_oi_neg_per_factor_analyze$direction = 'negative'

geneset_oi_analyze = rbind(geneset_oi_pos_per_factor_analyze, geneset_oi_neg_per_factor_analyze)
geneset_oi_analyze$fraction = as.numeric(top_variable_fraction)

geneset_oi_per_fraction = geneset_oi_analyze


## Calculate amount of top variables for the selected fraction
amount_geneset_oi_type = geneset_oi_analyze %>% group_by(type, view, variable) %>% dplyr::count()
amount_geneset_oi_type = merge(dimensions, amount_geneset_oi_type, all.x = TRUE)
amount_geneset_oi_type$fraction = as.numeric(top_variable_fraction)


geneset_oi_amounts =amount_geneset_oi_type


## Calculate total amount of features per type
features_per_type = feature_weights_long %>% group_by(type, view, variable) %>% dplyr::count()
colnames(features_per_type) = c('type', 'view', 'variable', 'amount_features')

## Merge top feature amounts and total amounts 
geneset_oi_amounts = merge(  geneset_oi_amounts,features_per_type, all.x = TRUE)

##  Calculate percentages
geneset_oi_amounts$percentage = geneset_oi_amounts$n / geneset_oi_amounts$amount_features

# Set NA when there are no features for this dimension among top %
geneset_oi_amounts[is.na(geneset_oi_amounts)] = 0 

### Prepare heatmap data (based on transformed expression values)

### Subset data on the selected factor (in the configuration)

analysis_data_subset = geneset_oi_per_fraction[geneset_oi_per_fraction$variable %in%  factor_var,]

### Add direction of effect

analysis_data_subset$direction_sign = analysis_data_subset$direction
analysis_data_subset$direction_sign[analysis_data_subset$direction == 'positive'] = '+'
analysis_data_subset$direction_sign[analysis_data_subset$direction == 'negative'] = '-'

head(analysis_data_subset,2)

### Add cell-type/ type column

analysis_data_subset$cell_type = analysis_data_subset$type

### Concat direction + view for faceting on y -axis

analysis_data_subset$facet =  paste0(analysis_data_subset$direction_sign, analysis_data_subset$view)

head(analysis_data_subset,2)

## Adjust column names in input data

data_long$expression_value = data_long$value
data_long$variable_name = data_long$variable
data_long$variable = NULL
data_long$value = NULL

head(data_long,2)

### Merge factor values and expression values

heatmap_data = merge(data_long, analysis_data_subset, by.x = c('variable_name'), by.y = c('variable_name'))

nrow(heatmap_data)

head(heatmap_data,2)

### Add sample meta-data

heatmap_data = merge(heatmap_data, sample_data, by.x = 'sample_id', by.y = 'sample_id')

head(heatmap_data,2)

# Heatmap Plot

###Generates heatmap based on configurations

## Specify figure name

figure_name = "FIG05_Heatmap_Feature_Overview"

# Specific Text Descriptions (heatmap) for x and y-axis:
xlabel_heatmap = xlab('Samples') 
ylabel_heatmap = ylab('View')

# Specific Text Descriptions (factor value plot) for x and y-axis:
xlabel_barplot = xlab('Samples') 
ylabel_barplot = ylab('Absolute Factor Value')

## Choose columns for faceting according to parameters specified in config file

## specify faceting for heamtap
heatmap_data$facet2 = heatmap_data[,variable] # x-axis faceting
heatmap_data$facet = paste0(heatmap_data$direction_sign, heatmap_data$view) # y-axis faceting

## specify faceting for factor value plot
analysis_data_subset$facet = paste0(analysis_data_subset$direction_sign,analysis_data_subset$view)
analysis_data_subset$facet2 = 'Absolute Factor Value'

view_var

## filter data on a specific view if specified in the config file
if(view_var %in% unique(heatmap_data$view)){
    heatmap_data = heatmap_data[heatmap_data$view %in% view_var,]
    analysis_data_subset = analysis_data_subset[analysis_data_subset$view %in% view_var,]
    } else{view_var = ''}

head(heatmap_data,2)

# Plot factor values as barplot

options(repr.plot.width=10, repr.plot.height=25)
factor_value_all_bar = ggplot(data = analysis_data_subset[(analysis_data_subset$variable == factor_var),] , 
                              mapping = aes(x = gene,   y= abs(value)))+
                            plot_config_heatmap_supp  + 
                            geom_bar(stat="identity", width = 0.1) + 
                            coord_flip()+
                            xlabel_barplot +
                            ylabel_barplot +
                            theme( axis.text.y = element_text( hjust = 1, vjust =1))+
                            facet_grid(facet  ~ facet2, scales = 'free', space = 'free', switch = 'y') + theme(strip.background = element_rect(color="black", fill="white", linetype="solid"), strip.placement = "outside", strip.switch.pad.grid = unit(0.5, "cm"))+
                            ggtitle('')


# Example of plot
#factor_value_all_bar

### Generate Heatmap Plot

factor_var

options(repr.plot.width=width_par, repr.plot.height=height_par)
all_heatmap = ggplot(data = heatmap_data[(heatmap_data$variable == factor_var) ,], # TBD whether to filter on sample subset
                     mapping = aes(sample_id, gene.x,  fill= expression_value)) +
    scale_fill_gradient2(low="#1D2ED8", high="#D8911D", midpoint = 0) + 
    geom_tile() + 
    xlabel_heatmap +
    ylabel_heatmap +
    plot_config_heatmap_supp  + theme(axis.text.x = element_blank()) + 
    facet_grid(facet  ~facet2, scales = 'free', space = 'free',  switch = "y") + theme(strip.background = element_rect(color="black", fill="white", linetype="solid"), strip.placement = "outside", strip.switch.pad.grid = unit(0.4, "cm"), strip.text.y.left = element_text(angle=0))+
    theme( axis.text.y = element_text( hjust = 1, vjust =1)) + 
    ggtitle(paste0('Normalized Expression Values of top ',as.numeric(top_variable_fraction) *100 *2, '% features'))

    

# Example heatmap plot
#all_heatmap

### Combined factor and heatamp plots and save the plot

legend =  get_legend(all_heatmap)

options(repr.plot.width=width_par, repr.plot.height=height_par)
combined = ggarrange(all_heatmap + theme(legend.position = 'none'),
         factor_value_all_bar + 
          theme(axis.text.y = element_blank(),axis.ticks.y = element_blank(),axis.title.y = element_blank(), strip.background.y = element_blank() , strip.text.y = element_blank()),  
         nrow=1, widths = c(4,1), align = 'h')
combined = annotate_figure(combined, right = legend)

pdf(paste0('figures/05_figures/', figure_name,view_var, '_', factor_var, '_',  top_variable_fraction,  mofa_name, '.pdf'), width =width_par, height =height_par)
print(combined)
dev.off()

### Inform about finalized execution

popup_function_pos('05_Downstream_Investigate_Features_Heatmap: Execution Finished')

Sys.sleep(20)
popup_function_info('05_Downstream_Investigate_Features_Heatmap')
