==> scripts/07_Compare_Models.ipynb <==
### Script to compare the results from different MOFA runs

#############################################
# Prerequisites - Load Libraries

source('MS1_Functions.r')

### Inform about execution start
popup_function_pos('07_Compare_Models: Execution Started')

source('MS0_Libraries.r')

source('MS2_Plot_Config.r')

###############################################
# Preqrequisites Configurations & Parameters

### Load the parameters that are set via the configuration files

### Load configurations file
global_configs = read.csv('configurations/Data_Configs.csv', sep = ',')

head(global_configs,2)

data_path = global_configs$value[global_configs$parameter == 'data_path']

data_path

result_path = global_configs$value[global_configs$parameter == 'result_path']

result_path

### Load the configuration file that specifies which MOFA models should be compared

comparison_configs = read.csv('configurations/07_Comparison_Configs.csv', sep = ',')

head(comparison_configs,2)

### Generate the result data directory if it does not exist yet
if(!file.exists(paste0(result_path, '07_results'))){
    dir.create(file.path(paste0(result_path, '07_results')))
    }

# Load Data 

## MOFA Models

### Load the trained MOFA Models

model_list = list()

comparison_configs$mofa_result_name

for(i in comparison_configs$mofa_result_name){
    model_name = paste0("03_MOFA_MODEL_", i, '.hdf5') # added number of factors in review to test different 
    outfile = file.path( paste0(result_path, '/03_results/',  model_name) )
    model_list[[as.character(i)]] <- load_model(outfile, verbose = TRUE)
    }

names(model_list)

#str(model_list)

print(length(model_list))

# Downstream Analysis of generated models

## Compare total explained variance

### Compare the explained variance per view between the models and plot the figure

lapply(model_list, function(x){
    mean(x@cache$variance_explained$r2_total[[1]])
    })
### 20 Factor model used: 35.7461964551804   

### compare explained variance per view

var_explained_per_view = lapply(model_list, function(x){
    x@cache[["variance_explained"]]$r2_total
    })

var_explained_per_view_data = data.frame()

for(i in names(var_explained_per_view)){
    data = var_explained_per_view[[i]]$group1
    data = as.data.frame(data)
    data$view = rownames(data)
    data$view = str_replace_all(data$view, ' ', '')
    data$view = str_replace_all(data$view, '\\.', '')
    data$configuration = i
    
    var_explained_per_view_data = rbind(data, var_explained_per_view_data)
    }

head(var_explained_per_view_data,2)

unique(var_explained_per_view_data$configuration)

# Specifies the figure name
figure_name = paste0("FIG07_Variance_Model_Comparison")

# Specifies the sizes of the plot
width_par = 6
height_par =4

plot = ggplot(var_explained_per_view_data, aes(x=view, y=data, fill=configuration, order = configuration)) + 
    geom_bar(position="dodge", stat="identity") + ylab('amount_explained_variance') + plot_config + scale_fill_okabe_ito() + coord_flip() + ylab('Total percentage of explained variance')
  # TBD: improve coloring + selection

#plot
pdf(paste0('figures/07_figures/', figure_name, '.pdf'), width =width_par, height =height_par)
print( plot)
dev.off()

## Compare factor weights of samples

## Compare the sample factor weights by correlating them (it is necessary that the factors match between the different models)

factor_weights = lapply(model_list, function(x){
    get_factors(x, factors = 'all')$group1
    })

factor_weight_data = data.frame()

for(i in names(factor_weights)){
    data = factor_weights[[i]]
    data = as.data.frame(data)
    data$configuration = i
    data$sample = rownames(data)
    data = melt(data)
    ## filter on relevant factors
    data = data[data$variable %in% unlist(str_split(comparison_configs$compare_factors[comparison_configs$mofa_result_name == i], ',')),]
    
    
    factor_weight_data = rbind(data, factor_weight_data)
    }

factor_weight_data$configuration = paste0(factor_weight_data$variable, '_', factor_weight_data$configuration)

head(factor_weight_data,2)

unique(factor_weight_data$variable)

cor_data = factor_weight_data %>% dcast(sample ~ configuration, value.var = 'value')

head(cor_data,2)

cor_data$sample = NULL
correlation = cor(cor_data, use = 'pairwise.complete.obs')

head(correlation,2)

## Specify the name of the plot
figure_name = paste0("FIG07_Factor_Correlations")

# Specify the sizes of the plot
width_par = 8
height_par =5

options(repr.plot.width=40, repr.plot.height=20)
pdf(paste0('figures/07_figures/', figure_name, '.pdf'), width =width_par, height =height_par)
corrplot(correlation, method = 'number',  type = 'upper', col = COL2('BrBG'), tl.col = 'black', tl.cex = 0.6,   number.cex = 0.4, tl.offset = 0.2, diag = FALSE) + plot_config + theme(text = element_text(size = 5))
dev.off()

## Generate a correlation table and save it 

correlation = melt(correlation)

correlation = correlation[correlation$Var1 != correlation$Var2,]  # remove diagonal entries

colnames(correlation) = c('Factor_Config1', 'Factor_Config2', 'Pearson_Correlation')

head(correlation ,2)

## Save result
write.csv(correlation, paste0(result_path, '/07_results/07_Factor_Correlations',  '.csv'), row.names = FALSE)

## Compare feature weights

## Compare the feature factor weights by correlating them (it is necessary that the features match between the different models)

weights = lapply(model_list, function(x){
    get_weights(x, views = "all", factors = "all")
    })

weight_data = data.frame()

for(j in names(weights)){
    weights_factor_amount = weights[[j]]

    for (i in names(weights_factor_amount)){
        data = data.frame(weights_factor_amount[[i]])
        data$view = i
        data$configuration = j
        data$feature = rownames(data)
        
        data = melt(data)
        # filter on relevant factors for comparison
        data = data[data$variable %in% unlist(str_split(comparison_configs$compare_factors[comparison_configs$mofa_result_name == j], ',')),]
        
        weight_data = rbind(weight_data,data)
        }
    }

head(weight_data,2)

weight_data$configuration = paste0(weight_data$variable, '_',weight_data$configuration)
weight_data$view_feature = paste0(weight_data$view, '__', weight_data$feature)
weight_data$view_feature =  str_replace_all(weight_data$view_feature, ' ', '')
weight_data$view_feature = str_replace_all(weight_data$view_feature, '\\.', '')

head(weight_data,2)

cor_data = weight_data %>% dcast(view_feature ~ configuration, value.var = 'value')

head(cor_data,2)

cor_data$view_feature = NULL
correlation = cor(cor_data, use = 'pairwise.complete.obs')

head(correlation,2)

figure_name = paste0("FIG07_Feature_Correlations")

# Sizes of the plot
width_par = 8
height_par =5

options(repr.plot.width=40, repr.plot.height=20)
pdf(paste0('figures/07_figures/', figure_name, '.pdf'), width =width_par, height =height_par)
corrplot(correlation, method = 'number',  type = 'lower', col = COL2('BrBG'), tl.col = 'black', tl.cex = 0.6, tl.srt = 90, number.cex = 0.4, tl.offset = 0.2, diag = FALSE) + plot_config + theme(text = element_text(size = 5))
dev.off()

## Generate Table

correlation = melt(correlation)

correlation = correlation[correlation$Var1 != correlation$Var2,]  # remove diagonal entries

colnames(correlation) = c('Factor_Config1', 'Factor_Config2', 'Pearson_Correlation')

head(correlation ,2)

## Save result
write.csv(correlation, paste0(result_path, '/07_results/07_Feature_Correlations',  '.csv'), row.names = FALSE)

### Inform about execution finalization
popup_function_pos('07_Compare_Models: Execution Finished')

Sys.sleep(20)
popup_function_info('07_Compare_Models')
